make conformed mappings have similar style
